refactor: 스크리닝 api 분리
