+++ src/apis/screening/screening.ts
@@ -1,16 +1,7 @@
 import {api} from '@/apis';
 import {ResponseDTO} from '@/models/common/responseDTO';
-import {
-  IScreeningBodyRequest,
-  IScreeningListOptionBodyRequest,
-  IScreeningListSearchBodyRequest,
-} from '@/models/screening/request';
-import {
-  IScreeningDetailContent,
-  TScreeningListResponse,
-  TScreeningReviewContent,
-  TWeekScreeningResponse,
-} from '@/models/screening/response';
+import {IScreeningBodyRequest} from '@/models/screening/request';
+import {TWeekScreeningResponse} from '@/models/screening/response';
 
 // 스크리닝 업로드 함수
 export const postScreening = async (body: IScreeningBodyRequest) => {
@@ -18,7 +9,7 @@ export const postScreening = async (body: IScreeningBodyRequest) => {
   return res.data;
 };
 
-// 이번주 스크리닝 함수
+// 이번주 스크리닝 가져오는 함수
 export const getWeekScreening = async (): Promise<
   ResponseDTO<TWeekScreeningResponse>
 > => {
@@ -26,7 +17,7 @@ export const getWeekScreening = async (): Promise<
   return res.data;
 };
 
-// 가장 최근 실시간 3개 스크리닝 함수
+// 가장 최근 실시간 3개 스크리닝 가져오는 함수
 export const getRecentScreening = async (): Promise<
   ResponseDTO<TWeekScreeningResponse>
 > => {
@@ -34,46 +25,10 @@ export const getRecentScreening = async (): Promise<
   return res.data;
 };
 
-// 댓글 많은 스크리닝 함수
+// 댓글 많은 스크리닝 가져오는 함수
 export const getMostCommentScreening = async (): Promise<
   ResponseDTO<TWeekScreeningResponse>
 > => {
   const res = await api.get('/screening/most-reviewed');
-  return res.data;
-};
-
-// 스크리닝 목록 시간 옵션 고르는 함수
-export const getTimeOptionScreeningList = async (
-  body: IScreeningListOptionBodyRequest,
-): Promise<ResponseDTO<TScreeningListResponse>> => {
-  const res = await api.get(
-    `/screening/screenings/search-by-date?category=${body.category}&page=${body.page}&size=${body.size}&sortBy=${body.sortBy}`,
-  );
-  return res.data;
-};
-
-// 스크리닝 목록 검색 함수
-export const getSearchScreeningList = async (
-  body: IScreeningListSearchBodyRequest,
-): Promise<ResponseDTO<TScreeningListResponse>> => {
-  const res = await api.get(
-    `/screening/screenings/search?title=${body.title}&category=${body.category}&page=${body.page}&size=${body.size}`,
-  );
-  return res.data;
-};
-
-// 스크리닝 디테일 id 정보 함수
-export const getScreeningDetailContent = async (
-  id: number,
-): Promise<ResponseDTO<IScreeningDetailContent>> => {
-  const res = await api.get(`/screening/${id}`);
-  return res.data;
-};
-
-// 특정 스크리닝에 리뷰 리스트 가져오는 함수
-export const getScreeningDetailReview = async (
-  id: number,
-): Promise<ResponseDTO<TScreeningReviewContent[]>> => {
-  const res = await api.get(`/screening/screening-review/${id}`);
   return res.data;
 };
